backend/requirements/base.txt
# Core Django
Django==4.2.17
djangorestframework==3.14.0
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Validation & Filtering
django-filter==23.3
drf-yasg==1.21.7

# AI
google-generativeai==0.3.2

# Cache & Queue
django-redis==5.3.0
celery==5.3.4
django-celery-beat==2.7.0

# Channels (WebSocket)
channels==4.0.0
channels-redis==4.1.0

# Files & Media
Pillow==10.1.0

# Utils
python-dateutil==2.8.2
pytz==2023.3